feat: set up dependency injection for application and infrastructure services
