+++ backend/src/Eventik.Application/Eventik.Application.csproj
@@ -13,9 +13,7 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Folder Include="DTOs\" />
       <Folder Include="Mappings\" />
-      <Folder Include="Services\" />
     </ItemGroup>
 
 </Project>
