+++ backend/src/Eventik.Core/Eventik.Core.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <Folder Include="Services\" />
+      <PackageReference Include="FluentResults" Version="3.16.0" />
     </ItemGroup>
 
 </Project>
